# app/Http/Resources/ReservationReportResource.php
<?php

namespace App\Http\Resources;

use Illuminate\Http\Request;
use Illuminate\Http\Resources\Json\JsonResource;

class ReservationReportResource extends JsonResource
{
    /**
     * Transform the resource into an array.
     *
     * @return array<string, mixed>
     */
    public function toArray(Request $request): array
    {
        return [
            'space_name' => $this->space_name,
            'date' => $this->date,
            'reservation_count' => $this->reservation_count,
        ];
    }
}