# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Reservations API - Documentation",
        "description": "Detailed API description for the space reservation system, by José González.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/reports/reservations-by-space": {
            "get": {
                "tags": [
                    "Reports (Admin)"
                ],
                "summary": "Get reservation report by space",
                "operationId": "649c4e29c74ea9601c0ab00feba55c05",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "e.g. 2025-01-01",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "e.g. 2025-12-31",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "space_id",
                        "in": "query",
                        "description": "Optional Space ID to filter",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservation report data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ReservationReportResource"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reports/active-users": {
            "get": {
                "tags": [
                    "Reports (Admin)"
                ],
                "summary": "Get a report of active users (most reservations)",
                "operationId": "87b081102d9a8ed734ac0774704c98e4",
                "responses": {
                    "200": {
                        "description": "List of active users sorted by reservation count",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ActiveUserReportResource"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ValidationError": {
                "title": "Validation Error",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "ReservationReportResource": {
                "title": "Reservation Report Resource",
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ValidationError"
                    },
                    {
                        "properties": {
                            "space_name": {
                                "type": "string",
                                "example": "Sala de Juntas A"
                            },
                            "date": {
                                "type": "string",
                                "format": "date",
                                "example": "2025-08-01"
                            },
                            "reservation_count": {
                                "type": "integer",
                                "example": 5
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "ActiveUserReportResource": {
                "title": "Active User Report Resource",
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ValidationError"
                    },
                    {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Usuario Admin"
                            },
                            "email": {
                                "type": "string",
                                "format": "email",
                                "example": "admin@example.com"
                            },
                            "role": {
                                "type": "string",
                                "example": "admin"
                            },
                            "reservations_count": {
                                "type": "integer",
                                "example": 12
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}